feat: add watch history feature
Introduce a history_service that tracks each time a user watches a video
(distinct from deduped video views), with endpoints to list, add, delete a
single entry, and clear all history. History is cleaned up on user/video
deletion via event consumers.

- Add history_service schema, sqlc queries, service, and error definitions
- Add GET /history, POST /history/videos/:video_id, DELETE /history/:entry_id,
  and DELETE /history handlers with tests
- Wire the service into the handler, routes, and main
- Use int instead of int64 at the app boundary, casting to sqlc int64 at the
  query layer (also applied to watchlater/playlist IDs and follow counts)
- Extract lock keys into local variables above each lock acquisition

+++ internals/handlers/watchlater.go
@@ -9,7 +9,7 @@ import (
 )
 
 func (me *Handler) HandleGetWatchlaters(c fiber.Ctx) error {
-	pr, err := parsePaginatedRequest[int64](c)
+	pr, err := parsePaginatedRequest[int](c)
 	if err != nil {
 		return err
 	}
